move : common/service 폴더 util로 이름 변경
